refactor: move Exhibit Card to common folder
