+++ app/components/common/cards/small/exhibit/index.tsx
